Refactor and modularize code, remove old api 0.X
